Move image helper into resource
